refactor: fix casing consistency for all API files and imports to prevent case-sensitive build errors
