+++ frontend/src/features/student/community/faqs/api/faqsApi.ts
