+++ frontend/src/features/student/community/resources/hooks/useResourceList.ts
@@ -2,7 +2,7 @@
 
 import { useCallback, useEffect, useState } from "react";
 import { ResourceListItemDto, PageMeta } from "../api/types";
-import { fetchResourcesList } from "../api/ResourcesApi";
+import { fetchResourcesList } from "../api/resourcesApi";
 
 
 const defaultMeta: PageMeta = {
@@ -58,7 +58,7 @@ export function useResourcesList() {
       items,
       meta,   // ✅ 항상 PageMeta
       page,
-      size, 
+      size,
       keyword,
       loading,
       error,
