+++ frontend/src/features/student/study-space/spaces-rentals/api/rentalsApi.ts
