+++ frontend/src/features/student/study-space/spaces-rentals/hooks/useRental.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { useCallback, useEffect, useMemo, useState } from "react";
-import { rentalsApi } from "../api/RentalsApi";
+import { rentalsApi } from "../api/rentalsApi";
 import type {
   AuthMeDto,
   RentalDto,
