+++ frontend/src/features/study-space/spaces-rentals/api/rentalsApi.ts
