+++ frontend/src/features/study-space/spaces/components/detail/SpacesDetailPage.client.tsx
@@ -4,7 +4,7 @@ import { useEffect, useMemo, useRef, useState } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
 import styles from "./SpacesDetailPage.module.css";
 
-import { spacesApi } from "../../api/SpacesApi";
+import { spacesApi } from "../../api/spacesApi";
 import type { SpaceDetailDto } from "../../api/types";
 import { Button } from "@/components/button";
 import SpacesRoomModal from "../modal/SpacesRoomModal.client";
