+++ frontend/src/features/study-space/spaces/components/list/SpacesPage.client.tsx
@@ -4,7 +4,7 @@ import { useEffect, useRef, useState } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
 
 import { PaginationSimple, useListQuery } from "@/components/pagination";
-import { spacesApi } from "../../api/SpacesApi";
+import { spacesApi } from "../../api/spacesApi";
 import type { PageMeta, SpaceListItemDto } from "../../api/types";
 import { SpacesTable } from "./SpacesTable";
 import styles from "./SpacesPage.module.css";
@@ -91,15 +91,15 @@ export default function SpacesPageClient() {
 
       <div className={styles.bottomRow}>
         <div className={styles.paginationWrap}>
-            <PaginationSimple page={page} totalPages={totalPages} onChange={(p) => setPage(p)} />
+          <PaginationSimple page={page} totalPages={totalPages} onChange={(p) => setPage(p)} />
         </div>
 
         <div className={styles.createWrap}>
-            <Button className={styles.createBtn} variant="primary" onClick={onClickCreate}>
+          <Button className={styles.createBtn} variant="primary" onClick={onClickCreate}>
             등록
-            </Button>
+          </Button>
         </div>
-        </div>
+      </div>
     </div>
   );
 }
